switch digital-twin dbcp and folder names
- [CORE-599]
