+++ dapps/assets/tsconfig.json
@@ -8,7 +8,7 @@
   ],
   "references": [
     {
-      "path": "../lib.digital-twin"
+      "path": "../digital-twin.lib"
     }
   ]
 }
