+++ dapps/assets/webpack.config.js
@@ -25,6 +25,6 @@ module.exports = require('../../scripts/dapp/webpack.config')(
   true,
   false,
   getExternals({
-    '@evan.network/lib.digital-twin': '@evan.network/lib.digital-twin',
+    '@evan.network/digital-twin.lib': '@evan.network/digital-twin.lib',
   })
 );
