+++ dapps/digital-twin.detail/.npmignore
