+++ dapps/digital-twin.detail/.npmrc
