+++ dapps/digital-twin.detail/src/components/data/DetailData.ts
