+++ dapps/digital-twin.detail/src/components/overview/DetailOverview.ts
