+++ dapps/digital-twin.detail/src/components/registry.ts
