+++ dapps/digital-twin.detail/src/i18n/de.json
