+++ dapps/digital-twin.detail/src/i18n/translations.ts
