+++ dapps/digital-twin.lib/src/DAppContract.ts
