+++ dapps/digital-twin.lib/webpack.config.js
